massive increase in doc-string coverage
