+++ ssms/basic_simulators/__init__.py
@@ -1,5 +1,6 @@
 from . import boundary_functions
 from . import drift_functions
 from . import simulator
+from . import theta_processor
 
-__all__ = ["boundary_functions", "drift_functions", "simulator", "utils"]
+__all__ = ["boundary_functions", "drift_functions", "simulator", "theta_processor"]
